# openapi/gogo.v1.yaml
openapi: 3.1.0
info:
  title: GoGo API
  version: 1.0.0-alpha.1
  description: >
    GoGo BFF public contract. Conventions: base path /v1, JSON UTF-8, ISO-8601
    UTC timestamps, money as integer minor units + currency, cursor pagination,
    Idempotency-Key on retryable mutations, optimistic concurrency via resource
    version/ETag. DTOs return facts, not composed sentences.
servers:
  - url: /v1
tags:
  - name: system
  - name: sessions
  - name: rooms
  - name: preferences
  - name: places
  - name: suggestions
  - name: plans
  - name: me
  - name: cms

paths:
  /health:
    get:
      operationId: getHealth
      tags: [system]
      summary: Liveness probe
      security: []
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]

  /auth/register:
    post:
      operationId: register
      tags: [sessions]
      summary: Register a user account (optionally claiming a guest session)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, minLength: 10, maxLength: 128 }
                displayName: { type: string, minLength: 1, maxLength: 50 }
                claimGuestToken:
                  type: string
                  description: Opaque guest token to re-parent guest history (FR-AUTH-003).
      responses:
        '201':
          description: Registered; web clients also receive HttpOnly session cookies.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenGrant' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Registration could not be completed (never confirms account existence)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429': { $ref: '#/components/responses/RateLimited' }

  /auth/login:
    post:
      operationId: login
      tags: [sessions]
      summary: Login with email + password (enumeration-safe, lockout on abuse)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '201':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenGrant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }

  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [sessions]
      summary: Rotate a refresh token or renew a guest access token
      description: >
        Refresh tokens are single-use; presenting a superseded token revokes
        the whole session family. Web clients may rely on the HttpOnly refresh
        cookie instead of the body field.
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
                guestToken: { type: string }
      responses:
        '201':
          description: New token grant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenGrant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }

  /sessions/guest:
    post:
      operationId: createGuestSession
      tags: [sessions]
      summary: Join a room as a guest with a display name only (FR-AUTH-002)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roomCode, displayName]
              properties:
                roomCode: { type: string, minLength: 6, maxLength: 64 }
                displayName: { type: string, minLength: 1, maxLength: 50 }
      responses:
        '201':
          description: Guest session created; access token is scoped to this room only.
          content:
            application/json:
              schema:
                type: object
                required: [guestSessionId, roomId, accessToken, guestToken]
                properties:
                  guestSessionId: { type: string, format: uuid }
                  roomId: { type: string, format: uuid }
                  accessToken: { type: string }
                  guestToken:
                    type: string
                    description: Long-lived opaque credential for re-entry and claim.
        '404': { $ref: '#/components/responses/NotFound' }
        '410':
          description: Room expired or no longer joinable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '429': { $ref: '#/components/responses/RateLimited' }

  /sessions/current:
    delete:
      operationId: endCurrentSession
      tags: [sessions]
      summary: Logout — revoke the current session (or all devices)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                allDevices: { type: boolean, default: false }
      responses:
        '200':
          description: Session revoked; cookies cleared.
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Current actor profile facts
      responses:
        '200':
          description: Actor profile (shape depends on actorType)
          content:
            application/json:
              schema:
                type: object
                required: [actorType, id]
                properties:
                  actorType: { type: string, enum: [user, guest] }
                  id: { type: string, format: uuid }
                  displayName: { type: string }
                  email: { type: string }
                  locale: { type: string }
                  roomId: { type: string, format: uuid }
                  expiresAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }

  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a couple/group room with constraint v1 and optional seed places
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, decisionMode, participantCount, constraint]
              properties:
                type: { type: string, enum: [couple, group] }
                decisionMode:
                  type: string
                  enum: [match, vote, host]
                  description: couple allows match|host; group allows vote|host.
                participantCount: { type: integer, minimum: 2, maximum: 20 }
                title: { type: string, maxLength: 80 }
                scheduledDate: { type: string, format: date-time }
                constraint: { $ref: '#/components/schemas/RoomConstraintInput' }
                seedPlaceIds:
                  type: array
                  maxItems: 10
                  items: { type: string, format: uuid }
      responses:
        '201':
          description: Room created (creator becomes host)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }

  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Room summary facts (members only; share code is host-only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Room summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  /rooms/{id}/constraints:
    patch:
      operationId: updateRoomConstraints
      tags: [rooms]
      summary: 'Host-only: new constraint version; marks scores/plans stale'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RoomConstraintInput'
                - type: object
                  required: [expectedConstraintVersion]
                  properties:
                    expectedConstraintVersion: { type: integer, minimum: 1 }
                    participantCount: { type: integer, minimum: 2, maximum: 20 }
      responses:
        '200':
          description: Updated summary with bumped constraintVersion
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }

  /rooms/{id}/status:
    patch:
      operationId: transitionRoom
      tags: [rooms]
      summary: 'Host-only: validated room state transition (SRS §7.2)'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [draft, collecting, matching, ready, active, completed, cancelled, expired]
      responses:
        '200':
          description: Transitioned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoomSummary' }
        '409': { $ref: '#/components/responses/Conflict' }

  /rooms/{id}/members:
    get:
      operationId: listRoomMembers
      tags: [rooms]
      summary: Member progress facts (never other members' preference selections)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RoomMember' }

  /rooms/{id}/members/{memberId}:
    delete:
      operationId: removeRoomMember
      tags: [rooms]
      summary: 'Host-only: remove a member (revokes guest session if guest)'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: memberId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Removed }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  /rooms/{id}/invites:
    post:
      operationId: createRoomInvite
      tags: [rooms]
      summary: 'Host-only: create a high-entropy expiring invite (code shown once)'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                maxUses: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '201':
          description: Invite created — plaintext code returned exactly once, only hash stored.
          content:
            application/json:
              schema:
                type: object
                required: [inviteId, code, expiresAt]
                properties:
                  inviteId: { type: string, format: uuid }
                  code: { type: string }
                  expiresAt: { type: string, format: date-time }
                  maxUses: { type: integer }
    get:
      operationId: listRoomInvites
      tags: [rooms]
      summary: 'Host-only: invite metadata (never codes)'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Invites }

  /rooms/{id}/invites/{inviteId}:
    delete:
      operationId: revokeRoomInvite
      tags: [rooms]
      summary: 'Host-only: revoke an invite'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: inviteId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Revoked }

  /rooms/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Authenticated user joins via invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteCode]
              properties:
                inviteCode: { type: string }
      responses:
        '201':
          description: Joined (idempotent for existing members)
          content:
            application/json:
              schema:
                type: object
                properties:
                  roomId: { type: string, format: uuid }
                  memberId: { type: string, format: uuid }
                  role: { type: string, enum: [host, member] }
        '410':
          description: Invite expired/revoked/spent or room not joinable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  /rooms/join/guest:
    post:
      operationId: joinRoomAsGuest
      tags: [rooms]
      summary: Guest joins via invite code without an account (FR-AUTH-002)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteCode, displayName]
              properties:
                inviteCode: { type: string }
                displayName: { type: string, minLength: 1, maxLength: 50 }
      responses:
        '201':
          description: Guest session bound to the invite's room
          content:
            application/json:
              schema:
                type: object
                properties:
                  guestSessionId: { type: string, format: uuid }
                  roomId: { type: string, format: uuid }
                  accessToken: { type: string }
                  guestToken: { type: string }
        '410':
          description: Invite expired/revoked/spent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  /rooms/{id}/seed-places:
    post:
      operationId: addRoomSeedPlaces
      tags: [rooms]
      summary: 'Host-only: attach verified places to the room (FR-ROOM-010)'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [placeIds]
              properties:
                placeIds:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items: { type: string, format: uuid }
      responses:
        '200': { description: Updated seed list }
        '400': { $ref: '#/components/responses/BadRequest' }

  /rooms/{id}/seed-places/{placeId}:
    delete:
      operationId: removeRoomSeedPlace
      tags: [rooms]
      summary: 'Host-only: remove a seed place'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: placeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated seed list }

  /rooms/{roomId}/preferences/me:
    get:
      operationId: getMyPreferences
      tags: [preferences]
      summary: Caller's own preference draft (private to the member)
      parameters:
        - { name: roomId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Own preferences with version }
    put:
      operationId: saveMyPreferences
      tags: [preferences]
      summary: Autosave with optimistic concurrency (expectedVersion)
      parameters:
        - { name: roomId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selections, expectedVersion]
              properties:
                selections:
                  type: object
                  additionalProperties:
                    type: array
                    items: { type: string }
                  description: Taxonomy kind → selected stable keys.
                weights:
                  type: object
                  additionalProperties: { type: number }
                expectedVersion: { type: integer, minimum: 0 }
      responses:
        '200': { description: 'Saved: { version, isDraft }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }

  /rooms/{roomId}/preferences/complete:
    post:
      operationId: completeMyPreferences
      tags: [preferences]
      summary: Mark own preferences complete; room auto-moves to matching when everyone is done
      parameters:
        - { name: roomId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: '{ completed, roomReadyForMatching }' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /taxonomies:
    get:
      operationId: listTaxonomies
      tags: [preferences]
      summary: Stable taxonomy keys + i18n labels (clients never hard-code ids)
      security: []
      parameters:
        - name: kinds
          in: query
          schema: { type: string }
          description: Comma-separated kind filter (mood,category,setting,dietary,accessibility,spending_style,suitability)
      responses:
        '200': { description: 'kinds → [{ key, sortOrder, labels }]' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: gogo_at
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Client-generated key for retryable mutations. Repeating a request with
        the same key returns the original result instead of re-applying it.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    RoomConstraintInput:
      type: object
      required: [budgetMode, budgetAmount]
      properties:
        originText: { type: string, maxLength: 200 }
        originLat: { type: number, minimum: -90, maximum: 90 }
        originLng: { type: number, minimum: -180, maximum: 180 }
        areaKey: { type: string }
        radiusM: { type: integer, minimum: 100, maximum: 50000 }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        budgetMode: { type: string, enum: [total, per_person] }
        budgetAmount:
          type: integer
          minimum: 0
          description: Integer minor units, interpreted per budgetMode.
        currency: { type: string, minLength: 3, maxLength: 3, default: VND }
        dietaryKeys:
          type: array
          items: { type: string }
        accessibilityKeys:
          type: array
          items: { type: string }
    RoomMember:
      type: object
      required: [id, displayName, role, selectionStatus, isGuest]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        role: { type: string, enum: [host, member] }
        selectionStatus: { type: string, enum: [pending, in_progress, completed] }
        isGuest: { type: boolean }
        joinedAt: { type: string, format: date-time }
    RoomSummary:
      type: object
      description: Facts only — audience copy is composed client-side from type/participantCount/budgetMode.
      required: [id, type, status, decisionMode, participantCount, constraintVersion]
      properties:
        id: { type: string, format: uuid }
        code:
          type: string
          description: Share code — present for the host only.
        type: { type: string, enum: [couple, group] }
        status:
          type: string
          enum: [draft, collecting, matching, ready, active, completed, cancelled, expired]
        decisionMode: { type: string, enum: [match, vote, host] }
        participantCount: { type: integer }
        constraintVersion: { type: integer }
        title: { type: string }
        scheduledDate: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        myMemberId: { type: string, format: uuid }
        myRole: { type: string, enum: [host, member] }
        constraints:
          allOf:
            - $ref: '#/components/schemas/RoomConstraintInput'
        members:
          type: array
          items: { $ref: '#/components/schemas/RoomMember' }
        seedPlaces:
          type: array
          items:
            type: object
            properties:
              placeId: { type: string, format: uuid }
              name: { type: string }
    TokenGrant:
      type: object
      required: [accessToken, expiresIn]
      properties:
        userId: { type: string, format: uuid }
        claimedRoomId: { type: string, format: uuid }
        accessToken: { type: string }
        refreshToken:
          type: string
          description: Absent for guest access renewal.
        roomId:
          type: string
          format: uuid
          description: Present for guest access renewal only.
        expiresIn:
          type: integer
          description: Access token TTL in seconds.
    ErrorEnvelope:
      type: object
      required: [code, message, field_errors, request_id, retryable]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
        field_errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
        request_id:
          type: string
        retryable:
          type: boolean
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          description: Integer minor units (e.g. VND has 0 decimals, amount = dong).
        currency:
          type: string
          minLength: 3
          maxLength: 3
  responses:
    BadRequest:
      description: Validation or business rule failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Actor lacks permission on this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Version/state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }

security:
  - bearerAuth: []